AlphaAgents/requirements.txt
# AlphaAgents Requirements
agno>=0.1.0
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
google-generativeai>=0.8.0

# Data and Analysis
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.0
pandas-datareader>=0.10.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.0.0

# Visualization (Optional - for backtesting charts)
# Uncomment to enable visualization features
# matplotlib>=3.7.0
# seaborn>=0.12.0
